Remove trailing slash from `/b` and `/o` URLs
Bucket [`list`][1] and [`insert`][2] operations have URL paths ending in
`/b`. Likewise, object [`list`][3] and [`insert`][4] operations have URL
paths ending in `/o`.

`googleCloudStorageR` was invoking these operations with a trailing
slash (`/b/` and `/o/`, respectively). This worked because the Google
Cloud Storage API seems to be lenient here. However, they still are
undocumented URLs that simply happen to work, which can lead to
unexpected issues. For example, it breaks compatibility with GCS
emulators that might only implement the *official* API.

This commit adjusts the request URLs for these operations so that they
match the documented ones (i.e., path ending in `/b` and `/o`, instead
of `/b/` and `/o/`).

[1]: https://cloud.google.com/storage/docs/json_api/v1/buckets/list
[2]: https://cloud.google.com/storage/docs/json_api/v1/buckets/insert
[3]: https://cloud.google.com/storage/docs/json_api/v1/objects/list
[4]: https://cloud.google.com/storage/docs/json_api/v1/objects/insert